Externalization of APIs so that Nautilus stays a resolver centric tool, and allows for easier maintenance and additions of web interfaces
